backend/requirements.txt
fastapi[standard]>=0.110.0
pydantic>=2.0.0
uvicorn>=0.20.0
pywebview>=5.0.0
pyinstaller>=6.4.0
